🎨 reorganise to parameterframe example
